Trained PPO and Recurrent PPO on ECG
trained PPO and Recurrent PPO for 10 million training timesteps on a 10-env VecEnv. PPO did show significant learning, but failed to exceed the reward expectations of the best 'non-communicative' stratefy. Recurrent PPO failed to learn entirely, which is understandable as it is an experimental model, and may need to be tuned to learn at all.